requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
socksio==1.0.0
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.19
starlette>=0.35.0
bcrypt>=4.0.0
playwright>=1.49.0
pycryptodome>=3.20.0
tos>=2.9.0
twilio>=9.0.0,<10
segno>=1.6.0
